+++ packages/editor/editor-core/src/plugins/media/ui/ResizableMediaSingle/index.tsx
@@ -131,24 +131,12 @@ export class ResizableMediaSingle extends React.Component<Props, State> {
 
   calcOffsetLeft() {
     let offsetLeft = 0;
-
     if (this.wrapper && this.insideInlineLike) {
-      let currentNode: HTMLElement | null = this.wrapper;
-      const pm = this.props.view.dom as HTMLElement;
-
-      while (
-        currentNode &&
-        currentNode.parentElement &&
-        !currentNode.parentElement.classList.contains('ProseMirror') &&
-        currentNode !== document.body
-      ) {
-        offsetLeft += currentNode.offsetLeft;
-        currentNode = currentNode.parentElement;
-      }
-
-      offsetLeft -= pm.offsetLeft;
+      const currentNode: HTMLElement = this.wrapper;
+      const boundingRect = currentNode.getBoundingClientRect();
+      const pmRect = this.props.view.dom.getBoundingClientRect();
+      offsetLeft = boundingRect.left - pmRect.left;
     }
-
     return offsetLeft;
   }
 
@@ -203,7 +191,6 @@ export class ResizableMediaSingle extends React.Component<Props, State> {
         snapPoints.push(fullWidthPoint);
       }
     }
-
     return snapPoints;
   }
 
@@ -261,14 +248,20 @@ export class ResizableMediaSingle extends React.Component<Props, State> {
       pxWidth = wideWidth > containerWidth ? lineLength : wideWidth;
     } else if (layout === 'full-width') {
       pxWidth = containerWidth - akEditorBreakoutPadding;
-    } else if (pctWidth && origWidth && origHeight) {
+    } else if (pctWidth && origWidth && origHeight && pctWidth < 100) {
       pxWidth = Math.ceil(
         calcPxFromPct(pctWidth / 100, lineLength || containerWidth),
       );
     } else if (layout === 'center') {
       pxWidth = Math.min(origWidth, lineLength);
     } else if (alignmentLayouts.indexOf(layout) !== -1) {
-      pxWidth = Math.min(origWidth / 2, lineLength);
+      const halfLineLength = Math.ceil(lineLength / 2);
+
+      if (origWidth <= halfLineLength) {
+        pxWidth = origWidth;
+      } else {
+        pxWidth = halfLineLength;
+      }
     }
 
     // scale, keeping aspect ratio