+++ packages/editor/renderer/src/ui/MediaCard.tsx
@@ -1,10 +1,13 @@
 import * as React from 'react';
 import { Component } from 'react';
+
+import { filter, ADFEntity } from '@atlaskit/adf-utils';
 import {
   CardAppearance,
   CardDimensions,
   Card,
-  CardView,
+  CardLoading,
+  CardError,
   CardOnClickCallback,
   ImageResizeMode,
 } from '@atlaskit/media-card';
@@ -12,6 +15,7 @@ import { MediaType } from '@atlaskit/adf-schema';
 import {
   FileIdentifier,
   ExternalImageIdentifier,
+  Identifier,
 } from '@atlaskit/media-client';
 import {
   withImageLoader,
@@ -22,6 +26,7 @@ import {
   ImageLoaderState,
 } from '@atlaskit/editor-common';
 import { RendererAppearance } from './Renderer';
+import { RendererContext } from '../react';
 
 export interface MediaCardProps {
   id?: string;
@@ -41,22 +46,87 @@ export interface MediaCardProps {
   imageStatus?: ImageStatus;
   disableOverlay?: boolean;
   useInlinePlayer?: boolean;
+  rendererContext?: RendererContext;
 }
 
+const mediaIdentifierMap: Map<string, Identifier> = new Map();
+
+export const getListOfIdentifiersFromDoc = (doc?: ADFEntity): Identifier[] => {
+  if (!doc) {
+    return [];
+  }
+  return filter(doc, node => node.type === 'media').reduce(
+    (identifierList: Identifier[], mediaNode) => {
+      if (mediaNode.attrs) {
+        const { type, url: dataURI, id } = mediaNode.attrs;
+
+        if (type === 'file' && id) {
+          identifierList.push({
+            mediaItemType: 'file',
+            id,
+          });
+        } else if (type === 'external' && dataURI) {
+          identifierList.push({
+            mediaItemType: 'external-image',
+            dataURI,
+            name: dataURI,
+          });
+        }
+      }
+      return identifierList;
+    },
+    [],
+  );
+};
+
 export class MediaCardInternal extends Component<MediaCardProps> {
+  componentDidMount() {
+    const {
+      rendererContext,
+      mediaProvider,
+      id,
+      url,
+      collection: collectionName,
+    } = this.props;
+
+    if (!mediaProvider) {
+      return;
+    }
+
+    const nodeIsInCache =
+      (id && mediaIdentifierMap.has(id)) ||
+      (url && mediaIdentifierMap.has(url));
+    if (rendererContext && rendererContext.adDoc && !nodeIsInCache) {
+      getListOfIdentifiersFromDoc(rendererContext.adDoc).forEach(identifier => {
+        if (identifier.mediaItemType === 'file') {
+          mediaIdentifierMap.set(identifier.id as string, {
+            ...identifier,
+            collectionName,
+          });
+        } else if (identifier.mediaItemType === 'external-image') {
+          mediaIdentifierMap.set(identifier.dataURI as string, identifier);
+        }
+      });
+    }
+  }
+
+  componentWillUnmount() {
+    const { id, url: dataURI } = this.props;
+
+    if (id) {
+      mediaIdentifierMap.delete(id);
+    } else if (dataURI) {
+      mediaIdentifierMap.delete(dataURI);
+    }
+  }
+
   private renderLoadingCard = () => {
     const { cardDimensions } = this.props;
 
-    return (
-      <CardView
-        status="loading"
-        mediaItemType="file"
-        dimensions={cardDimensions}
-      />
-    );
+    return <CardLoading dimensions={cardDimensions} />;
   };
 
-  private renderExternal() {
+  private renderExternal(shouldOpenMediaViewer: boolean) {
     const {
       cardDimensions,
       resizeMode,
@@ -83,6 +153,10 @@ export class MediaCardInternal extends Component<MediaCardProps> {
         appearance={appearance}
         resizeMode={resizeMode}
         disableOverlay={disableOverlay}
+        shouldOpenMediaViewer={shouldOpenMediaViewer}
+        mediaViewerDataSource={{
+          list: Array.from(mediaIdentifierMap.values()),
+        }}
       />
     );
   }
@@ -137,7 +211,7 @@ export class MediaCardInternal extends Component<MediaCardProps> {
     const shouldOpenMediaViewer = !isMobile && !onCardClick;
 
     if (type === 'external') {
-      return this.renderExternal();
+      return this.renderExternal(shouldOpenMediaViewer);
     }
 
     if (type === 'link') {
@@ -149,13 +223,7 @@ export class MediaCardInternal extends Component<MediaCardProps> {
     }
 
     if (!id || type !== 'file') {
-      return (
-        <CardView
-          status="error"
-          mediaItemType={type}
-          dimensions={cardDimensions}
-        />
-      );
+      return <CardError dimensions={cardDimensions} />;
     }
 
     const identifier: FileIdentifier = {
@@ -175,6 +243,9 @@ export class MediaCardInternal extends Component<MediaCardProps> {
         disableOverlay={disableOverlay}
         useInlinePlayer={isInlinePlayer}
         shouldOpenMediaViewer={shouldOpenMediaViewer}
+        mediaViewerDataSource={{
+          list: Array.from(mediaIdentifierMap.values()),
+        }}
       />
     );
   }