+++ packages/elements/emoji/src/__tests__/unit/api/media/SiteEmojiResourceSpec.ts
@@ -1,7 +1,7 @@
 jest.mock('@atlaskit/media-client');
 import 'es6-promise/auto'; // 'whatwg-fetch' needs a Promise polyfill
 
-import * as fetchMock from 'fetch-mock/src/client';
+import fetchMock from 'fetch-mock/src/client';
 import * as sinon from 'sinon';
 import { waitUntil } from '@atlaskit/util-common-test';
 import { FileState, MediaClient } from '@atlaskit/media-client';