+++ packages/media/media-viewer/src/__tests__/unit/newgen/header.spec.tsx
@@ -335,7 +335,7 @@ describe('<Header />', () => {
       el: ReactWrapper<any, any>,
       enabled: boolean,
     ) => {
-      expect(el.find({ type: 'button', isDisabled: !enabled })).toHaveLength(1);
+      expect(el.find({ isDisabled: !enabled }).find('button')).toHaveLength(1);
       expect(el.find(DownloadIcon)).toHaveLength(1);
     };