+++ packages/media/media-viewer/src/newgen/download.tsx
@@ -9,7 +9,7 @@ import {
   Identifier,
   isExternalImageIdentifier,
 } from '@atlaskit/media-client';
-import Button from '@atlaskit/button';
+import { MediaButton } from '@atlaskit/media-ui';
 import { withAnalyticsEvents } from '@atlaskit/analytics-next';
 import DownloadIcon from '@atlaskit/icon/glyph/download';
 import { DownloadButtonWrapper } from './styled';
@@ -29,7 +29,7 @@ export const DownloadButton: any = withAnalyticsEvents({
     const ev = createEvent(props.analyticsPayload);
     ev.fire(channel);
   },
-})(Button);
+})(MediaButton);
 
 export const createItemDownloader = (
   file: FileState,
@@ -98,7 +98,7 @@ export const ToolbarDownloadButton = (props: ToolbarDownloadButtonProps) => {
 };
 
 export const DisabledToolbarDownloadButton = (
-  <Button
+  <MediaButton
     appearance={'toolbar' as any}
     isDisabled={true}
     iconBefore={downloadIcon}