+++ packages/media/media-viewer/src/newgen/viewers/audio.tsx
@@ -29,6 +29,8 @@ export type Props = Readonly<{
   previewCount: number;
   featureFlags?: MediaViewerFeatureFlags;
   showControls?: () => void;
+  onCanPlay?: () => void;
+  onError?: () => void;
 }>;
 
 export type State = BaseState<string> & {
@@ -75,7 +77,7 @@ export class AudioViewer extends BaseViewer<string, Props, State> {
   };
 
   protected renderSuccessful(src: string) {
-    const { showControls, previewCount } = this.props;
+    const { showControls, previewCount, onCanPlay, onError } = this.props;
 
     const useCustomAudioPlayer = !isIE();
     const isAutoPlay = previewCount === 0;
@@ -89,6 +91,8 @@ export class AudioViewer extends BaseViewer<string, Props, State> {
             src={src}
             isShortcutEnabled={true}
             showControls={showControls}
+            onCanPlay={onCanPlay}
+            onError={onError}
           />
         </CustomAudioPlayerWrapper>
       </AudioPlayer>