+++ packages/elements/share/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @atlaskit/share
 
+## 0.3.8
+- [patch] [95293c5550](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/95293c5550):
+
+  - Added documentation and consolidated example
+
+## 0.3.7
+- Updated dependencies [1e826b2966](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e826b2966):
+  - @atlaskit/docs@7.0.2
+  - @atlaskit/analytics-next@4.0.3
+  - @atlaskit/field-text-area@5.0.3
+  - @atlaskit/form@5.2.5
+  - @atlaskit/icon@16.0.8
+  - @atlaskit/inline-dialog@10.0.3
+  - @atlaskit/section-message@2.0.2
+  - @atlaskit/select@8.0.5
+  - @atlaskit/theme@8.1.6
+  - @atlaskit/toggle@6.0.3
+  - @atlaskit/tooltip@13.0.3
+  - @atlaskit/button@12.0.0
+
+## 0.3.6
+- [patch] [e03dea5f5d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e03dea5f5d):
+
+  - FS-3792 do no call loadUser on empty query
+
+## 0.3.5
+- [patch] [77c2d7bb2b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/77c2d7bb2b):
+
+  - Added flags for successful share
+
+## 0.3.4
+- [patch] [0f4109e919](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0f4109e919):
+
+  - FS-3743 remove loading message from user picker
+
+## 0.3.3
+- [patch] [66512e9026](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/66512e9026):
+
+  - FS-3764 move dependencies to peer dependencies
+
+## 0.3.2
+- [patch] [c68b454ba9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c68b454ba9):
+
+  - fixed unclear share panel state upon successful share
+
+## 0.3.1
+- [patch] [ddfc158dfb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ddfc158dfb):
+
+  - Removed unused buttonStyle prop
+
+## 0.3.0
+- [minor] [b617f099aa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b617f099aa):
+
+  - Limited length of comment messages.
+
+## 0.2.10
+- [patch] [69c72e07ba](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/69c72e07ba):
+
+  - Fixed share button off position on IE11
+
 ## 0.2.9
 - [patch] [fcdae04b8c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fcdae04b8c):