+++ packages/people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @atlaskit/profilecard
 
+## 9.0.1
+- Updated dependencies [1e826b2966](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e826b2966):
+  - @atlaskit/docs@7.0.2
+  - @atlaskit/avatar@15.0.3
+  - @atlaskit/dynamic-table@11.0.2
+  - @atlaskit/icon@16.0.8
+  - @atlaskit/popper@0.4.2
+  - @atlaskit/portal@0.3.1
+  - @atlaskit/spinner@10.0.5
+  - @atlaskit/theme@8.1.6
+  - @atlaskit/button@12.0.0
+
+## 9.0.0
+- [major] [bfca144ea5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bfca144ea5):
+
+  - Refactored internal components. Moved to Typescript. Updated named exports. Updated type definitions.
+
 ## 8.0.2
 - Updated dependencies [9d5cc39394](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9d5cc39394):
   - @atlaskit/docs@7.0.1